requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
av==17.0.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
ctranslate2==4.7.1
decorator==4.4.2
distro==1.9.0
edge-tts==7.2.8
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
groq==1.1.1
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyboard==0.13.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.1
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.4.3
onnxruntime==1.24.4
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytweening==1.2.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
streamlit-autorefresh==1.0.1
streamlit-option-menu==0.4.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
yarl==1.23.0
